requirements.txt
# Core Framework
streamlit==1.31.0

# LLM and Embeddings
anthropic==0.25.0
google-generativeai==0.3.2

# Document Processing
pypdf==4.0.1
python-docx==1.1.0

# Text Processing
tiktoken==0.6.0

# Utilities
python-dotenv==1.0.1